+++ brq/producer.py
@@ -143,116 +143,3 @@ class Producer(DeferOperator):
         defer_key = self.get_deferred_key(function_name)
         dead_key = self.get_dead_message_key(function_name)
         await self.redis.delete(stream_name, defer_key, dead_key)
-
-    async def remove_deferred_job(
-        self,
-        function_name: str,
-        job: Job,
-    ):
-        """
-        Remove specific deferred job
-
-        Args:
-            function_name (str): function name
-            job (Job): job to be removed
-        """
-        defer_key = self.get_deferred_key(function_name)
-        delete_nums = await self.redis.zrem(defer_key, job.to_redis())
-        logger.info(f"Removed {delete_nums} deferred jobs")
-
-    async def get_defer_timestamp_ms(
-        self,
-        defer_until: datetime = None,
-        defer_hours: int = 0,
-        defer_minutes: int = 0,
-        defer_seconds: int = 0,
-    ):
-        if not any(
-            [
-                defer_until,
-                defer_hours,
-                defer_minutes,
-                defer_seconds,
-            ]
-        ):
-            return None
-
-        if defer_until:
-            logger.debug(f"Using defer_until, ignore defer_hours, defer_minutes, defer_seconds")
-            defer_until = defer_until.timestamp()
-
-        else:
-            defer_until = (
-                await self.get_current_timestamp(self.redis)
-                + defer_hours * 60 * 60
-                + defer_minutes * 60
-                + defer_seconds
-            )
-
-        return defer_until * 1000
-
-    async def get_deferred_jobs(self, function_name: str) -> dict[datetime:Job]:
-        """
-        Get all deferred jobs
-
-        Args:
-            function_name (str): function name
-
-        Returns:
-            dict[datetime:Job]: deferred jobs,
-                key is the datetime when the job expected to be executed
-                value is the job
-        """
-        defer_key = self.get_deferred_key(function_name)
-        return {
-            datetime.fromtimestamp(float(element[1]) / 1000): Job.from_redis(element[0])
-            for element in await self.redis.zrangebyscore(
-                defer_key, "-inf", "+inf", withscores=True
-            )
-        }
-
-    async def get_dead_messages(self, function_name: str) -> dict[datetime:Job]:
-        """
-        Get all dead messages
-
-        Args:
-            function_name (str): function name
-
-        Returns:
-            dict[datetime:Job]: dead messages,
-                key is the datetime when the job to be moved to dead message
-                value is the job
-        """
-        dead_key = self.get_dead_message_key(function_name)
-        return {
-            datetime.fromtimestamp(float(element[1]) / 1000): Job.from_redis(element[0])
-            for element in await self.redis.zrangebyscore(dead_key, "-inf", "+inf", withscores=True)
-        }
-
-    async def count_deferred_jobs(self, function_name: str):
-        defer_key = self.get_deferred_key(function_name)
-        return await self.redis.zcard(defer_key)
-
-    async def count_stream(self, function_name: str):
-        """
-        Count stream length
-
-        If Consumer's `delete_messgae_after_process` if False, will include already processed messages
-        """
-        stream_name = self.get_stream_name(function_name)
-        return await self.redis.xlen(stream_name)
-
-    async def count_unacked_jobs(self, function_name: str, group_name: str = "default-workers"):
-        """
-        Count unacked jobs in group
-
-        Args:
-            function_name (str): function name
-            group_name (str, optional): group name. Defaults to "default-workers". Should be the same as Consumer's `group_name`
-        """
-        stream_name = self.get_stream_name(function_name)
-        return await self.redis.xpending(stream_name, group_name)
-
-    async def count_dead_messages(self, function_name: str):
-        dead_key = self.get_dead_message_key(function_name)
-        return await self.redis.zcard(dead_key)
